catch exception by reference and drop unused name
Catching by reference avoids the object slicing that will happen when a child of
std::exception (std::invalid_argument or std::out_of_range in this case) is
thrown. This is not so important in this case as we're about to exit anyway, but
it's a little cleaner to avoid leaking memory.