Fix SIGSEGV crash when LANG not detected
'getenv()' returns NULL if the value passed is not found in the
enviroment, and because of the assumed behaviour as it stands,
this may crash -- we're trying to cast a NULL value to std::string
which of course will not work.

We can remedy this by checking for the NULL value and setting a
fallback value if needed.
y